% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_anns}
\alias{gene_anns}
\title{Gene annotations}
\format{
A data frame with \code{GENE_NAME}, \code{GENE_TYPE}, and further annotation
columns (description, GO terms, synonyms), one row per gene.
}
\usage{
gene_anns
}
\description{
Gene annotations
}
\keyword{datasets}
