% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gene_boxplots}
\alias{plot_gene_boxplots}
\title{Boxplots of the final model's signature genes by true class}
\usage{
plot_gene_boxplots(fit, n_genes = NULL)
}
\arguments{
\item{fit}{A \code{gotnerve_fit} from \code{\link[=train_model]{train_model()}}.}

\item{n_genes}{Number of genes (ranked by |final coefficient|) to plot, or
\code{NULL} for all genes with a non-zero final coefficient.}
}
\value{
A ggplot object (faceted by gene).
}
\description{
Boxplots of the final model's signature genes by true class
}
