% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_risk_distribution}
\alias{plot_risk_distribution}
\title{Predicted-probability boxplot by true class}
\usage{
plot_risk_distribution(fit, dataset = c("train", "test"))
}
\arguments{
\item{fit}{A \code{gotnerve_fit} from \code{\link[=train_model]{train_model()}}.}

\item{dataset}{\code{"train"} or \code{"test"}.}
}
\value{
A ggplot object.
}
\description{
Predicted-probability boxplot by true class
}
