% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotNeRve-package.R
\docType{package}
\name{gotNeRve-package}
\alias{gotNeRve-package}
\title{gotNeRve: Predict Stromal Tumor-Associated Nerve (STaN) Abundance from Gene Expression}
\description{
Implements a gene expression-based classifier for predicting high or low stromal tumor-associated nerve (STaN) abundance in non-muscle-invasive bladder cancer. The bundled elastic-net model uses a 12-gene transcriptomic signature developed from bulk RNA-sequencing of tumors with matched image-based STaN quantification. Includes gotNeRve() for scoring raw counts, CPM, or log2-CPM gene expression data.
}
\author{
\strong{Maintainer}: Cailin Deiter \email{cailin.deiter@cuanschutz.edu}

Authors:
\itemize{
  \item Cailin Deiter \email{cailin.deiter@cuanschutz.edu}
}

}
\keyword{internal}
